New option for setting the return-path (-rp)
A new optional option -rp allows setting the return-path of the sent
e-mail. It works by using the given address when issuing the MAIL FROM
command to the smtp server. Smtp servers use this address in the
Return-Path header field.

This feature can be used to specify where bounces are sent.